+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ Changelog
 * Fix bug in :meth:`pykallisto.KallistoFrame.compute_fold_change` method.
 * Add new method :meth:`pyvcf.call` and new command :command:`vcf-call`.
 * Combine optional arguments ``bam`` and ``fn`` into single positional argument ``bams`` for :meth:`pycov.CovFrame.from_bam` method. The same goes for :command:`bam-depth` command (combine ``--bam`` and ``--fn`` into ``bams``).
+* Combine optional arguments ``bed`` and ``region`` into single optional argument ``regions`` for :meth:`pycov.CovFrame.from_bam` method. The same goes for :command:`bam-depth` command (combine ``--bed`` and ``--region`` into ``--regions``).
+* Update :meth:`pycov.CovFrame.from_bam` method and :command:`bam-depth` command to automatically handle the 'chr' string.
 
 0.30.0 (2022-02-05)
 -------------------